Revert "mysql"

+++ src/Moneyman.Domain/ITransaction.cs
@@ -6,8 +6,7 @@ namespace Moneyman.Domain
     public class BaseTransaction : Entity
     {
         public string Name {get; set;}
-        public int Amount {get; set;}
-
+        public decimal Amount {get; set;}
         public bool Active { get; set; }
 
         public DateTime StartDate { get; set; }
